# minor uop speedups [pr]