--- a/test/test_uops.py
+++ b/test/test_uops.py
@@ -3,7 +3,7 @@ import unittest, math
 import numpy as np
 from tinygrad.shape.shapetracker import ShapeTracker
 from tinygrad.tensor import Tensor, _to_np_dtype
-from tinygrad.helpers import CI, DEBUG, getenv, Context
+from tinygrad.helpers import CI, DEBUG, getenv, Context, Timing
 from tinygrad.dtype import dtypes, DType
 from tinygrad.device import Buffer, Device
 from tinygrad.ops import Ops, UOp, UPat, KernelInfo, exec_alu, spec # noqa F401
@@ -463,12 +463,16 @@ class TestUPatHelpers(unittest.TestCase):
       test_upat = UPat(Ops.CONST, dtypes.bool)
       self.assertEqual(test_upat.location[0].split("/")[-1], __file__.replace("\\", "/").split("/")[-1])
 
-class TestUopsImmutable(unittest.TestCase):
+class TestUopsObject(unittest.TestCase):
   @unittest.expectedFailure
   def test_immutable(self):
     const_4 = UOp.const(dtypes.int, 4)
     with self.assertRaises(Exception):
       const_4.arg = 5
+
+  def test_timing(self):
+    with Timing("create 10k uops:"): ret = [UOp(Ops.CONST, dtypes.int, arg=10000000+i) for i in range(10000)]
+    assert len(ret) == 10000
 
 if __name__ == '__main__':
   unittest.main(verbosity=2)