757 ms -> 710 ms

```
light@blak:~/tinygrad$ python3 test/external/external_benchmark_schedule.py
***** model tensor in     19.95 ms
***** model schedule in   59.47 ms
***** model opts(31) in  144.02 ms
***** model lower in      12.10 ms
***** model rewrite in   358.41 ms
***** model linearize in 120.16 ms
19675
all 715.89 ms
light@blak:~/tinygrad$ python3 test/external/external_benchmark_schedule.py
***** model tensor in     19.11 ms
***** model schedule in   57.05 ms
***** model opts(31) in  142.60 ms
***** model lower in      11.98 ms
***** model rewrite in   357.81 ms
***** model linearize in 119.30 ms
19675
all 709.55 ms
light@blak:~/tinygrad$ git checkout master
Switched to branch 'master'
Your branch is up to date with 'origin/master'.
light@blak:~/tinygrad$ python3 test/external/external_benchmark_schedule.py
***** model tensor in     19.45 ms
***** model schedule in   65.05 ms
***** model opts(31) in  149.90 ms
***** model lower in      13.44 ms
***** model rewrite in   386.07 ms
***** model linearize in 121.00 ms
19675
all 756.63 ms
light@blak:~/tinygrad$ python3 test/external/external_benchmark_schedule.py
***** model tensor in     19.01 ms
***** model schedule in   64.08 ms
***** model opts(31) in  151.05 ms
***** model lower in      13.47 ms
***** model rewrite in   388.45 ms
***** model linearize in 119.82 ms
19675
all 757.52 ms
```